make sure routes have the correct schema when they come out of the DB
