+++ deployables/app/app/routes/account/edit.spec.ts
@@ -1,26 +1,12 @@
 import { render } from '@/test-utils'
 import { screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import {
-  dbClient,
-  getAccount,
-  getDefaultRoute,
-  setDefaultRoute,
-} from '@zodiac/db'
+import { dbClient, getAccount } from '@zodiac/db'
 import {
   accountFactory,
-  routeFactory,
   tenantFactory,
   userFactory,
-  walletFactory,
 } from '@zodiac/db/test-utils'
-import {
-  createMockEoaAccount,
-  createMockRoute,
-  createMockStartingWaypoint,
-  createMockWaypoints,
-  randomAddress,
-} from '@zodiac/test-utils'
 import { href } from 'react-router'
 import { queryInitiators, queryRoutes } from 'ser-kit'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
@@ -85,174 +71,6 @@ describe('Edit account', () => {
         'label',
         'New label',
       )
-    })
-  })
-
-  describe('Route', () => {
-    it('is auto-selects the first route', async () => {
-      const tenant = await tenantFactory.create()
-      const user = await userFactory.create(tenant)
-      const account = await accountFactory.create(tenant, user)
-      const wallet = await walletFactory.create(user, {
-        label: 'Test wallet',
-      })
-
-      mockQueryInitiators.mockResolvedValue([wallet.address])
-
-      const waypoints = createMockWaypoints({
-        end: true,
-      })
-
-      const newRoute = createMockRoute({ id: 'first', waypoints })
-
-      mockQueryRoutes.mockResolvedValue([newRoute])
-
-      const { waitForPendingLoaders, waitForPendingActions } = await render(
-        href('/account/:accountId', {
-          accountId: account.id,
-        }),
-        { tenant, user },
-      )
-
-      await userEvent.click(
-        await screen.findByRole('combobox', { name: 'Pilot Signer' }),
-      )
-      await userEvent.click(
-        await screen.findByRole('option', { name: 'Test wallet' }),
-      )
-
-      await waitForPendingLoaders()
-
-      await userEvent.click(await screen.findByRole('button', { name: 'Save' }))
-
-      await waitForPendingActions()
-
-      const activeRoute = await getDefaultRoute(
-        dbClient(),
-        tenant,
-        user,
-        account.id,
-      )
-
-      expect(activeRoute.route).toHaveProperty('waypoints', waypoints)
-    })
-
-    it('is possible to change the selected route', async () => {
-      const tenant = await tenantFactory.create()
-      const user = await userFactory.create(tenant)
-      const account = await accountFactory.create(tenant, user)
-      const wallet = await walletFactory.create(user, {
-        label: 'Test wallet',
-      })
-
-      mockQueryInitiators.mockResolvedValue([wallet.address])
-
-      const waypoints = createMockWaypoints({
-        end: true,
-        start: createMockStartingWaypoint(
-          createMockEoaAccount({ address: randomAddress() }),
-        ),
-      })
-
-      const firstRoute = createMockRoute({
-        id: 'first',
-        waypoints: createMockWaypoints({
-          end: true,
-        }),
-      })
-
-      const route = await routeFactory.create(account, wallet, {
-        waypoints: firstRoute.waypoints,
-      })
-
-      await setDefaultRoute(dbClient(), tenant, user, route)
-
-      const secondRoute = createMockRoute({ id: 'second', waypoints })
-
-      mockQueryRoutes.mockResolvedValue([firstRoute, secondRoute])
-
-      const { waitForPendingActions } = await render(
-        href('/account/:accountId', {
-          accountId: account.id,
-        }),
-        { tenant, user },
-      )
-
-      // click last route in radio select
-      await userEvent.click((await screen.findAllByRole('radio'))[1])
-
-      await userEvent.click(await screen.findByRole('button', { name: 'Save' }))
-
-      await waitForPendingActions()
-
-      const activeRoute = await getDefaultRoute(
-        dbClient(),
-        tenant,
-        user,
-        account.id,
-      )
-
-      expect(activeRoute.route).toHaveProperty('waypoints', waypoints)
-    })
-  })
-
-  describe('Routes', () => {
-    it('lists all routes', async () => {
-      const tenant = await tenantFactory.create()
-      const user = await userFactory.create(tenant)
-
-      const wallet = await walletFactory.create(user)
-      const account = await accountFactory.create(tenant, user)
-
-      await routeFactory.create(account, wallet, {
-        label: 'Route A',
-      })
-      await routeFactory.create(account, wallet, {
-        label: 'Route B',
-      })
-
-      await render(href('/account/:accountId', { accountId: account.id }), {
-        user,
-        tenant,
-        features: ['multiple-routes'],
-      })
-
-      expect(
-        await screen.findByRole('tab', { name: 'Route A' }),
-      ).toBeInTheDocument()
-      expect(
-        await screen.findByRole('tab', { name: 'Route B' }),
-      ).toBeInTheDocument()
-    })
-
-    it('is shows the initiator of the specified route', async () => {
-      const tenant = await tenantFactory.create()
-      const user = await userFactory.create(tenant)
-
-      const walletA = await walletFactory.create(user)
-      const walletB = await walletFactory.create(user)
-      const account = await accountFactory.create(tenant, user)
-
-      const routeA = await routeFactory.create(account, walletA, {
-        label: 'Route A',
-      })
-      await routeFactory.create(account, walletB, {
-        label: 'Route B',
-      })
-
-      await setDefaultRoute(dbClient(), tenant, user, routeA)
-
-      mockQueryInitiators.mockResolvedValue([walletA.address, walletB.address])
-
-      await render(href('/account/:accountId', { accountId: account.id }), {
-        user,
-        tenant,
-        features: ['multiple-routes'],
-      })
-
-      await userEvent.click(await screen.findByRole('tab', { name: 'Route B' }))
-
-      expect(await screen.findByText(walletB.label)).toBeInTheDocument()
     })
   })
 })
