+++ packages/db/src/access/routes/getRoute.ts
@@ -1,4 +1,5 @@
 import { invariant } from '@epic-web/invariant'
+import { waypointsSchema } from '@zodiac/schema'
 import type { UUID } from 'crypto'
 import type { DBClient } from '../../dbClient'
 
@@ -14,5 +15,5 @@ export const getRoute = async (db: DBClient, routeId: UUID) => {
 
   invariant(route != null, `Could not find route with id "${routeId}"`)
 
-  return route
+  return { ...route, waypoints: waypointsSchema.parse(route.waypoints) }
 }
